requirements.txt
flask~=2.2.2
sqlalchemy~=2.0.21
mysql~==2.
python-dotenv~=0.21.0
gunicorn